requirements.txt
fastapi==0.111.1
uvicorn[standard]==0.29.0
requests==2.32.3
celery==5.4.0
redis==5.0.7
geopy==2.4.1
python-dotenv==1.0.1
pytz==2024.1
eventlet==0.35.2
